refactor(user): refactor wechat classes
